requirements.txt
beaker-pyteal==0.4.1
py-algorand-sdk==1.20.0
pytest==7.2.0